move more components
